+++ src/app/Data/services/forecast-data/forecast-data.service.spec.ts
@@ -1,11 +1,11 @@
-import { Observable, of } from 'rxjs';
-import { TestBed } from '@angular/core/testing';
 import { HttpClient } from '@angular/common/http';
+import { Observable, of, throwError } from 'rxjs';
+import { TestBed } from '@angular/core/testing';
 
 import { ForecastData } from '../../../Domain/models/forecast-data.model';
 import { ForecastDataService } from './forecast-data.service';
 
-describe('forecastDataService', (): void => {
+describe('ForecastDataService', (): void => {
   let service: ForecastDataService;
   let httpClientSpy: { get: jasmine.Spy } = jasmine.createSpyObj(
     'HttpClient',
@@ -87,5 +87,60 @@ describe('forecastDataService', (): void => {
 
     // Assert
     result.subscribe((data: ForecastData) => expect(data).toBeNull());
+  });
+
+  it('should return data with correct structure and values when given a valid id', (): void => {
+    // Mock the http.get method to return a mock Observable
+    const mockResponse: ForecastData = { location: 'forest', data: [], yMaxValue: 100 };
+    // Arrange
+    httpClientSpy.get.and.returnValue(of(mockResponse));
+
+    // Act
+    const result: Observable<ForecastData> = service.getForestData(1);
+
+    // Assert that the result has the correct structure and values
+    result.subscribe((forecastData: ForecastData) => {
+      expect(forecastData.location).toBe('forest');
+      expect(forecastData.data).toEqual([]);
+      expect(forecastData.yMaxValue).toBe(100);
+    });
+  });
+
+  it('should handle JSON file not found', (): void => {
+    // Arrange
+    httpClientSpy.get.and.returnValue(throwError((): string => 'File not found'));
+
+    // Act
+    const result: Observable<ForecastData> = service.getForestData(132);
+
+    // Assert that the result is an Observable of error
+    expect(result).toEqual(jasmine.any(Observable));
+    result.subscribe({
+      next: (): void => {
+        fail('Expected error to be thrown');
+      },
+      error: (error): void => {
+        expect(error).toBe('File not found');
+      }
+    });
+  });
+
+  it('should handle JSON file with invalid format', (): void => {
+    // Arrange
+    httpClientSpy.get.and.returnValue(throwError((): string => 'Invalid JSON'));
+
+    // Act
+    const result: Observable<ForecastData> = service.getForestData(31);
+
+    // Assert that the result is an Observable of error
+    expect(result).toEqual(jasmine.any(Observable));
+    result.subscribe({
+      next: (): void => {
+        fail('Expected error to be thrown');
+      },
+      error: (error): void => {
+        expect(error).toBe('Invalid JSON');
+      }
+    });
   });
 });