+++ src/app/Data/services/forecast-data/forecast-data.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { map, Observable } from 'rxjs';
+import { Observable } from 'rxjs';
 
 import { environment } from '../../../../environments/environment';
 import { ForecastData } from '../../../Domain/models/forecast-data.model';
@@ -11,10 +11,6 @@ export class ForecastDataService {
   }
 
   getForestData(id: number): Observable<ForecastData> {
-    return this.http
-      .get<ForecastData>(`${environment.dataPath}data_${id}.json`)
-      .pipe(
-        map((forestData: ForecastData) => forestData)
-      );
+    return this.http.get<ForecastData>(`${environment.dataPath}data_${id}.json`);
   }
 }