+++ CHANGELOG.md
@@ -63,4 +63,6 @@ All notable changes to this project will be documented in this file.
 
 ## [0.5.0]
 
-- Tests for the forecastService.
+- Tests for the ForecastService.
+- Tests for the TestCaseService.
+- Tests for the UserService.